+++ src/main/java/org/jgoeres/adventofcode2019/Day09/RunDay09.java
@@ -7,32 +7,32 @@ public abstract class RunDay09 {
     static IntCodeProcessorService intCodeProcessorService = new IntCodeProcessorService(pathToInputs);
 
 
-    public static int problem9A() {
+    public static Long problem9A() {
         // Day 9A
         System.out.println("=== DAY 9A ===");
 
-        // Set up the processor in the air conditioner test configuration
-        intCodeProcessorService.setCpuInputValue(1);
+        // Set up the BOOST program in test mode
+        intCodeProcessorService.setCpuInputValue(1L);
         // Run it
         intCodeProcessorService.runToCompletion();
 
-        int result = intCodeProcessorService.getProgramOutput();
+        Long result = intCodeProcessorService.getProgramOutput();
         System.out.println("Day 9A: Program Output = " + result);
-//        Day 9A: Program Output = 9775037
-//        Time elapsed:	17 ms
+//        Day 9A: Program Output = 4080871669
+//        Time elapsed:	73 ms
         return result;
     }
 
-    public static int problem9B() {
+    public static Long problem9B() {
         // Day 9B
         System.out.println("=== DAY 9B ===");
 
         // Set the processor up for the thermal radiator controller test
         intCodeProcessorService.reset();
-        intCodeProcessorService.setCpuInputValue(5);
+        intCodeProcessorService.setCpuInputValue(5L);
         intCodeProcessorService.runToCompletion();
 
-        int result = intCodeProcessorService.getProgramOutput();
+        Long result = intCodeProcessorService.getProgramOutput();
 //        System.out.println("Day 9B: Program Output = " + result);
 //        Day 9B: Program Output = 15586959
 //        Time elapsed:	1 ms
